Reply with OK for unhandled events
When set to notify everything, github will send a lot of irrelevant events
that are not pushes or pings. Since we have already validated the security token
we know it's coming from GitHub, so there's no need to report a bad request:
we can just report OK and avoid the error spam on Github's side.